+++ kotlinx-uuid-datetime/api/kotlinx-uuid-datetime.api
@@ -1,4 +1,4 @@
-public final class kotlinx/uuid/datetime/DslKt {
+public final class app/softwork/uuid/datetime/DslKt {
 	public static final fun UUIDv7 (Lkotlinx/datetime/Instant;Lkotlin/random/Random;)Lkotlin/uuid/Uuid;
 	public static synthetic fun UUIDv7$default (Lkotlinx/datetime/Instant;Lkotlin/random/Random;ILjava/lang/Object;)Lkotlin/uuid/Uuid;
 	public static final fun getInstant (Lkotlin/uuid/Uuid;)Lkotlinx/datetime/Instant;
