+++ kotlinx-uuid-sqldelight/README.md
@@ -1,8 +1,8 @@
 # Module kotlinx-uuid-sqldelight
 
-SQLDelight uses column adapters for custom types, like this UUID.
-`kotlinx-uuid-sqldelight` provides two adapters, a `UUIDStringAdapter` for a `String` and a `ByteArrayAdapter` for
-a `ByteArray` representation respectively.
+SQLDelight uses column adapters for custom types, like [Uuid](kotlin.uuid.Uuid).
+`kotlinx-uuid-sqldelight` provides two adapters, a [UuidStringAdapter](app.softwork.uuid.sqldelight.UuidStringAdapter) for a [String] and a [UuidByteArrayAdapter](app.softwork.uuid.sqldelight.UuidByteArrayAdapter)` for
+a [ByteArray] representation respectively.
 
 ```kotlin
 dependencies {
