+++ kotlinx-uuid-sqldelight/src/commonMain/kotlin/app/softwork/uuid/sqldelight/UuidByteArrayAdapter.kt
@@ -1,0 +1,9 @@
+package app.softwork.uuid.sqldelight
+
+import app.cash.sqldelight.*
+import kotlin.uuid.Uuid
+
+public object UuidByteArrayAdapter : ColumnAdapter<Uuid, ByteArray> {
+    override fun decode(databaseValue: ByteArray): Uuid = Uuid.fromByteArray(databaseValue)
+    override fun encode(value: Uuid): ByteArray = value.toByteArray()
+}
